He creado Principal.java para probar mi programa y he tenido que cambiar la variable en el metodo factorial para que funcionase mejor. He detectado este fallo tras pasar el depurador

+++ com.entornos.FernandoJesusFernandezTrujillo.algoritmos/src/com/entornos/FernandoJesusFernandezTrujillo/algoritmos/algoritmos.java
@@ -27,7 +27,7 @@ public abstract class algoritmos {
 	 * @param numero
 	 * @return el factorial del numero introducido
 	 */
-	 public static int factorial( int numero ) {
+	 public static long factorial( int numero ) {
 	      int fact = 1;
 	      for( int i = 1; i <= numero; i++ ) {
 	         fact *= i;
